Misc cleanup and refactoring in CF... taking the liberty of just modifying 'my' code at will here.
git-svn-id: https://svn.apache.org/repos/asf/lucene/mahout/trunk@805747 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mahout/cf/taste/impl/recommender/FarthestNeighborClusterSimilarity.java
@@ -81,7 +81,7 @@ public final class FarthestNeighborClusterSimilarity implements ClusterSimilarit
     }
     // We skipped everything? well, at least try comparing the first Users to get some value
     if (leastSimilarity == Double.POSITIVE_INFINITY) {
-      return similarity.userSimilarity((Long) cluster1.iterator().next(), (Long) cluster2.iterator().next());
+      return similarity.userSimilarity(cluster1.iterator().next(), cluster2.iterator().next());
     }
     return leastSimilarity;
   }
